Move the retryable function to the retry module
